Allow mods to provide additional package loaders.
