+++ OpenRA.Game/FileSystem/BagFile.cs
@@ -11,117 +11,130 @@
 
 using System;
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Text;
 using OpenRA.FileFormats;
 using OpenRA.Primitives;
 
 namespace OpenRA.FileSystem
 {
-	public sealed class BagFile : IReadOnlyPackage
+	public class BagFileLoader : IPackageLoader
 	{
-		public string Name { get; private set; }
-		public IEnumerable<string> Contents { get { return index.Keys; } }
-
-		readonly Stream s;
-		readonly Dictionary<string, IdxEntry> index;
-
-		public BagFile(FileSystem context, string filename)
+		public sealed class BagFile : IReadOnlyPackage
 		{
-			Name = filename;
+			public string Name { get; private set; }
+			public IEnumerable<string> Contents { get { return index.Keys; } }
 
-			// A bag file is always accompanied with an .idx counterpart
-			// For example: audio.bag requires the audio.idx file
-			var indexFilename = Path.ChangeExtension(filename, ".idx");
+			readonly Stream s;
+			readonly Dictionary<string, IdxEntry> index;
 
-			// Build the index and dispose the stream, it is no longer needed after this
-			List<IdxEntry> entries;
-			using (var indexStream = context.Open(indexFilename))
-				entries = new IdxReader(indexStream).Entries;
-
-			index = entries.ToDictionaryWithConflictLog(x => x.Filename,
-				"{0} (bag format)".F(filename),
-				null, x => "(offs={0}, len={1})".F(x.Offset, x.Length));
-
-			s = context.Open(filename);
-		}
-
-		public Stream GetStream(string filename)
-		{
-			IdxEntry entry;
-			if (!index.TryGetValue(filename, out entry))
-				return null;
-
-			s.Seek(entry.Offset, SeekOrigin.Begin);
-
-			var waveHeaderMemoryStream = new MemoryStream();
-
-			var channels = (entry.Flags & 1) > 0 ? 2 : 1;
-
-			if ((entry.Flags & 2) > 0)
+			public BagFile(FileSystem context, string filename)
 			{
-				// PCM
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("RIFF"));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length + 36));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("WAVE"));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("fmt "));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(16));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)1));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)channels));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.SampleRate));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(2 * channels * entry.SampleRate));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)(2 * channels)));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)16));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("data"));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length));
+				Name = filename;
+
+				// A bag file is always accompanied with an .idx counterpart
+				// For example: audio.bag requires the audio.idx file
+				var indexFilename = Path.ChangeExtension(filename, ".idx");
+
+				// Build the index and dispose the stream, it is no longer needed after this
+				List<IdxEntry> entries;
+				using (var indexStream = context.Open(indexFilename))
+					entries = new IdxReader(indexStream).Entries;
+
+				index = entries.ToDictionaryWithConflictLog(x => x.Filename,
+					"{0} (bag format)".F(filename),
+					null, x => "(offs={0}, len={1})".F(x.Offset, x.Length));
+
+				s = context.Open(filename);
 			}
 
-			if ((entry.Flags & 8) > 0)
+			public Stream GetStream(string filename)
 			{
-				// IMA ADPCM
-				var samplesPerChunk = (2 * (entry.ChunkSize - 4)) + 1;
-				var bytesPerSec = (int)Math.Floor(((double)(2 * entry.ChunkSize) / samplesPerChunk) * ((double)entry.SampleRate / 2));
-				var chunkSize = entry.ChunkSize > entry.Length ? entry.Length : entry.ChunkSize;
+				IdxEntry entry;
+				if (!index.TryGetValue(filename, out entry))
+					return null;
 
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("RIFF"));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length + 52));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("WAVE"));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("fmt "));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(20));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)17));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)channels));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.SampleRate));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(bytesPerSec));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)chunkSize));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)4));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)2));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)samplesPerChunk));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("fact"));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(4));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(4 * entry.Length));
-				waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("data"));
-				waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length));
+				s.Seek(entry.Offset, SeekOrigin.Begin);
+
+				var waveHeaderMemoryStream = new MemoryStream();
+
+				var channels = (entry.Flags & 1) > 0 ? 2 : 1;
+
+				if ((entry.Flags & 2) > 0)
+				{
+					// PCM
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("RIFF"));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length + 36));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("WAVE"));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("fmt "));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(16));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)1));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)channels));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.SampleRate));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(2 * channels * entry.SampleRate));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)(2 * channels)));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)16));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("data"));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length));
+				}
+
+				if ((entry.Flags & 8) > 0)
+				{
+					// IMA ADPCM
+					var samplesPerChunk = (2 * (entry.ChunkSize - 4)) + 1;
+					var bytesPerSec = (int)Math.Floor(((double)(2 * entry.ChunkSize) / samplesPerChunk) * ((double)entry.SampleRate / 2));
+					var chunkSize = entry.ChunkSize > entry.Length ? entry.Length : entry.ChunkSize;
+
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("RIFF"));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length + 52));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("WAVE"));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("fmt "));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(20));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)17));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)channels));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.SampleRate));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(bytesPerSec));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)chunkSize));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)4));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)2));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes((short)samplesPerChunk));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("fact"));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(4));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(4 * entry.Length));
+					waveHeaderMemoryStream.Write(Encoding.ASCII.GetBytes("data"));
+					waveHeaderMemoryStream.Write(BitConverter.GetBytes(entry.Length));
+				}
+
+				waveHeaderMemoryStream.Seek(0, SeekOrigin.Begin);
+
+				// Construct a merged stream
+				var mergedStream = new MergedStream(waveHeaderMemoryStream, s);
+				mergedStream.SetLength(waveHeaderMemoryStream.Length + entry.Length);
+
+				return mergedStream;
 			}
 
-			waveHeaderMemoryStream.Seek(0, SeekOrigin.Begin);
+			public bool Contains(string filename)
+			{
+				return index.ContainsKey(filename);
+			}
 
-			// Construct a merged stream
-			var mergedStream = new MergedStream(waveHeaderMemoryStream, s);
-			mergedStream.SetLength(waveHeaderMemoryStream.Length + entry.Length);
-
-			return mergedStream;
+			public void Dispose()
+			{
+				s.Dispose();
+			}
 		}
 
-		public bool Contains(string filename)
+		public bool TryParsePackage(FileSystem context, string filename, out IReadOnlyPackage package)
 		{
-			return index.ContainsKey(filename);
-		}
+			if (!filename.EndsWith(".bag", StringComparison.InvariantCultureIgnoreCase))
+			{
+				package = null;
+				return false;
+			}
 
-		public void Dispose()
-		{
-			s.Dispose();
+			package = new BagFile(context, filename);
+			return true;
 		}
 	}
 }
