+++ OpenRA.Game/FileSystem/IdxEntry.cs
@@ -9,7 +9,6 @@
  */
 #endregion
 
-using System.Collections.Generic;
 using System.IO;
 
 namespace OpenRA.FileSystem
